[REFACTOR] Finalize ticket naming standardization across codebase
