+++ src/events/ticket_events.rs
@@ -1,58 +1,78 @@
+use crate::model::ticket::ticket::Ticket;
 use std::sync::{Arc, Mutex};
 use uuid::Uuid;
-use crate::model::tiket::ticket::Ticket;
+use std::fmt::Debug;
 
+/// Defines the possible ticket-related events
 #[derive(Clone, Debug)]
 pub enum TicketEvent {
     Created(Ticket),
     Updated(Ticket),
     Deleted(Uuid),
     Allocated { ticket_id: Uuid, quantity: u32 },
+    Purchased { ticket_id: Uuid, user_id: Uuid, quantity: u32, transaction_id: Uuid },
     SoldOut(Uuid),
+    Validated { ticket_id: Uuid, validator_id: Uuid },
 }
 
-pub trait TicketEventObserver: Send + Sync {
-    fn on_event(&self, event: TicketEvent);
+/// Interface for objects that can handle ticket events
+pub trait TicketEventObserver: Send + Sync + Debug {
+    fn on_event(&self, event: &TicketEvent);
 }
 
+/// Manages ticket events and observers
+#[derive(Debug)]
 pub struct TicketEventManager {
-    observers: Arc<Mutex<Vec<Arc<dyn TicketEventObserver>>>>,
+    observers: Mutex<Vec<Arc<dyn TicketEventObserver>>>,
 }
 
 impl TicketEventManager {
     pub fn new() -> Self {
         Self {
-            observers: Arc::new(Mutex::new(Vec::new())),
+            observers: Mutex::new(Vec::new()),
         }
     }
-    
+
+    /// Add an observer that will be notified of ticket events
     pub fn add_observer(&self, observer: Arc<dyn TicketEventObserver>) {
         let mut observers = self.observers.lock().unwrap();
         observers.push(observer);
     }
-    
+
+    /// Notify all registered observers about a ticket event
     pub fn notify_observers(&self, event: TicketEvent) {
         let observers = self.observers.lock().unwrap();
         for observer in observers.iter() {
-            observer.on_event(event.clone());
+            observer.on_event(&event);
         }
     }
 }
 
-// Example observer
-pub struct EmailNotifier;
+/// Sample observer that sends email notifications
+#[derive(Debug)]
+pub struct EmailNotifier {}
+
+impl EmailNotifier {
+    pub fn new() -> Self {
+        Self {}
+    }
+}
 
 impl TicketEventObserver for EmailNotifier {
-    fn on_event(&self, event: TicketEvent) {
+    fn on_event(&self, event: &TicketEvent) {
         match event {
-            TicketEvent::SoldOut(ticket_id) => {
-                println!("Sending email: Ticket {} is now sold out!", ticket_id);
-                // In a real system, this would call an email service
-            },
             TicketEvent::Created(ticket) => {
-                println!("Sending email: New ticket created for event {}", ticket.event_id);
+                println!("📧 Email: Ticket type '{}' for event {} has been created", 
+                    ticket.ticket_type, ticket.event_id);
             },
-            _ => {} // Handle other events as needed
+            TicketEvent::Purchased { ticket_id, user_id, quantity, .. } => {
+                println!("📧 Email: User {} has purchased {} tickets (ID: {})", 
+                    user_id, quantity, ticket_id);
+            },
+            TicketEvent::SoldOut(ticket_id) => {
+                println!("📧 Email: Ticket {} is now sold out!", ticket_id);
+            },
+            _ => {} // Other events don't trigger emails
         }
     }
 }
