+++ src/model/mod.rs
@@ -1,1 +1,3 @@
+// This file declares the ticket and transaction submodules
+pub mod ticket;
 pub mod transaction;
