+++ src/repository/adapters/postgres_adapter.rs
@@ -1,7 +1,7 @@
 use sqlx::{PgPool, Row};
 use uuid::Uuid;
-use crate::model::tiket::ticket::{Ticket, TicketStatus};
-use crate::repository::tiket::TicketRepository;
+use crate::model::ticket::ticket::{Ticket, TicketStatus};
+use crate::repository::ticket::TicketRepository;
 
 pub struct PostgresTicketRepository {
     pool: PgPool,
